commento inutile

+++ pydef/templates/struct_templates.c
@@ -2,7 +2,7 @@ typedef struct task_struct {
     int e_wc;
     struct task_struct* in_set[3];
     int in_set_count;
-    float deadline;
+    unsigned long int deadline;
     void (*function_pointer)();
     char isEnabled;
 } task_struct_t;
@@ -11,7 +11,7 @@ typedef struct app_struct {
     task_struct_t* app_tasks[5];
     task_struct_t* initial_task;
     int tasks_count;
-    float x_min;
+    unsigned long int x_min;
     char isActive;
 } app_struct_t;
